+++ plugins/benq/qscan_plugin.cpp
@@ -156,7 +156,7 @@ int  scan_benq::start_test(unsigned int itest, long ilba, int &speed)
 	}
 }
 
-int  scan_benq::scan_block(void *data, long *ilba)
+int  scan_benq::scan_block(void *data, uint32_t *ilba)
 {
 	int r=-1;
 	switch (test) {
