+++ plugins/generic/qscan_plugin.cpp
@@ -79,7 +79,7 @@ int  scan_generic::start_test(unsigned int itest, long ilba, int &speed) {
 	return r;
 }
 
-int  scan_generic::scan_block(void *data, long *ilba) {
+int  scan_generic::scan_block(void *data, uint32_t *ilba) {
 	int r=-1;
 	switch (test) {
 		case CHK_ERRC_CD:
