+++ plugins/generic/qscan_plugin.h
@@ -36,17 +36,17 @@ public:
     virtual int  errc_data();
     virtual int  check_test(unsigned int test);
     virtual int  start_test(unsigned int test, long slba, int &speed);
-    virtual int  scan_block(void* data,long* ilba);
+    virtual int  scan_block(void* data,uint32_t* ilba);
     virtual int  end_test();
 
     virtual const char* name() { return plugin_name; };
     virtual const char* desc() { return plugin_desc; };
 private:
-	long lba;
+	uint32_t lba;
 
 	int cmd_cd_errc_block(cd_errc *data);
 //	int c2calc(cd_errc *data);
-	int c2calc(unsigned char* buf, unsigned int lba, unsigned char sects);
+	int c2calc(unsigned char* buf, uint32_t lba, unsigned char sects);
 };
 
 #endif
