+++ plugins/pioneer/qscan_plugin.cpp
@@ -127,7 +127,7 @@ int  scan_pioneer::start_test(unsigned int itest, long ilba, int &speed)
 	return r;
 }
 
-int  scan_pioneer::scan_block(void *data, long *ilba)
+int  scan_pioneer::scan_block(void *data, uint32_t *ilba)
 {
 	int r=-1;
 	switch (test) {
