+++ plugins/plextor/qscan_plugin.cpp
@@ -195,7 +195,7 @@ int  scan_plextor::start_test(unsigned int itest, long ilba, int &speed)
 	return r;
 }
 
-int  scan_plextor::scan_block(void *data, long *ilba)
+int  scan_plextor::scan_block(void *data, uint32_t *ilba)
 {
 	int r=-1;
 	switch (test) {
