+++ plugins/tsst/qscan_plugin.h
@@ -36,13 +36,13 @@ public:
     virtual int  errc_data();
     virtual int  check_test(unsigned int test);
     virtual int  start_test(unsigned int test, long slba, int &speed);
-    virtual int  scan_block(void* data,long* ilba);
+    virtual int  scan_block(void* data,uint32_t* ilba);
     virtual int  end_test();
 
     virtual const char* name() { return plugin_name; };
     virtual const char* desc() { return plugin_desc; };
 private:
-	long lba;
+	uint32_t lba;
 
 // CD ERRC methods
 	int cmd_cd_errc_init();
